Added calendar views etc, checklist/button todo
