+++ public/js/todo.js
@@ -1,7 +1,6 @@
 const deleteBtn = document.querySelectorAll('.del')
-const deletePets = document.querySelectorAll('.delpet')
-const todoItem = document.querySelectorAll('span.not')
-const todoComplete = document.querySelectorAll('span.completed')
+const todoItem = document.querySelectorAll('label.not')
+const todoComplete = document.querySelectorAll('label.completed')
 const other = document.getElementById('otherTodo')
 const inputOther = document.getElementById('inputOtherTodo')
 
@@ -11,10 +10,6 @@ inputOther.addEventListener('change', () => {
 
 Array.from(deleteBtn).forEach((el)=>{
     el.addEventListener('click', deleteTodo)
-})
-
-Array.from(deletePets).forEach((el)=>{
-    el.addEventListener('click', deletePet)
 })
 
 Array.from(todoItem).forEach((el)=>{
@@ -33,23 +28,6 @@ async function deleteTodo(){
             headers: {'Content-type': 'application/json'},
             body: JSON.stringify({
                 'todoIdFromJSFile': todoId
-            })
-        })
-        const data = await response.json()
-        location.reload()
-    }catch(err){
-        console.error(err)
-    }
-}
-
-async function deletePet(){
-    const petId = this.parentNode.dataset.id
-    try{
-        const response = await fetch('pets/deletePet', {
-            method: 'delete',
-            headers: {'Content-type': 'application/json'},
-            body: JSON.stringify({
-                'petIdFromJSFile': petId
             })
         })
         const data = await response.json()
