+++ routes/calendar.js
@@ -1,0 +1,7 @@
+const express = require('express')
+const router = express.Router()
+const calendarController = require('../controllers/calendar') 
+
+router.get('/', calendarController.getCalendar)
+
+module.exports = router
