+++ views/dashboard.ejs
@@ -43,6 +43,9 @@
             <span class="visually-hidden">(current)</span>
           </li>
           <li class="nav-item">
+            <a class="nav-link" href="/calendar">Calendar</a>
+          </li>
+          <li class="nav-item">
             <a class="nav-link" href="/logout">Logout</a>
           </li>
         </ul>
@@ -79,5 +82,6 @@
       </div>
     </div>
 </div>
+<script src="js/pet.js"></script>
 </body>
 </html>
